+++ megamek/src/megamek/common/RefrigeratedCargoBay.java
@@ -25,7 +25,7 @@ public final class RefrigeratedCargoBay extends Bay {
      */
     private static final long serialVersionUID = 4161027191694822726L;
 
-    private float weight = 0;
+    private double weight = 0;
 
     /**
      * The default constructor is only for serialization.
@@ -48,7 +48,7 @@ public final class RefrigeratedCargoBay extends Bay {
      */
     public RefrigeratedCargoBay(double space, int doors, int bayNumber) {
         totalSpace = space * 0.87;
-        weight = (float) space;
+        weight = space;
         currentSpace = space * 0.87;
         this.doors = doors;
         this.bayNumber = bayNumber;
@@ -91,7 +91,7 @@ public final class RefrigeratedCargoBay extends Bay {
     }
 
     @Override
-    public float getWeight() {
+    public double getWeight() {
         return weight;
     }
